# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Users, Documentation",
        "description": "Esta API é utilizada para gerenciar um sistema de acessos em geral, oferecendo funcionalidades para autenticação de usuários, gerenciamento de contas, e administração de privilégios. Ela suporta três níveis de acesso: Administrador, Atendente e Cliente.\n  - Administradores podem acessar e modificar dados de qualquer usuário.\n  - Atendentes podem gerenciar contas, como update, deletar usuarios e ver uma lista completa de dados de cada usuario e acessar algumas funcionalidades restritas.\n  - Clientes têm acesso limitado às suas próprias contas.\n\n **Autenticação**\n - A API utiliza autenticação baseada em tokens **(Bearer Token via Sanctum)**. Todos os endpoints que requerem autenticação estão devidamente marcados com a configuração de segurança. O objetivo principal desta documentação é ajudar desenvolvedores a integrar suas aplicações com o sistema, fornecendo detalhes sobre requisições, respostas, e exemplos práticos de uso.\n\n  **Principais funcionalidades**:\n  - Registro, login e logout de usuários.\n  - Gestão de privilégios para controle de acesso.\n  - Operações CRUD em usuários (somente admins e atendentes).",
        "contact": {
            "email": "oliveirarafael22@outlook.com.br"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://api-users-rafael.up.railway.app/api/v1/",
...
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Efetua o login do usuario",
                "description": "Para Logar no sistema, caso nao tenha uma conta, Crie em Resgister, e com o Email e Senha substitua os campos, Clicando em **Try it out**, e Obtenha o token de autorização.",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Input do email do login",
                                        "type": "string",
                                        "example": "edward@teste.com"
                                    },
                                    "password": {
                                        "description": "Input da Senha da conta",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso. Após obter o token de autorização inclua-no no **Authorize**",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "000|pnQZ3VxhbHU49DWTycJOvZgLaKob6RB5Gn8zVtZsb9b7455a"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, caso tente acessar com algum email que nao exista, ou a senha incorreta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos obrigatórios não foram preenchidos. Caso um ou mais campos sejam omitidos, será exibida uma mensagem de erro.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de erro quando um campo obrigatório não é enviado.",
                                            "type": "string",
                                            "example": "O campo email é Obrigatorio"
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "Lista os erros específicos do campo.",
                                                        "type": "string",
                                                        "example": "O campo email é Obrigatorio"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza logout do usuário",
                "description": "Quando o usuario quiser sair do sistema, ele efetuará o Logout",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado. Faça login para continuar.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Realiza o cadastro de Usuarios.",
                "description": "Realiza o cadastro de usuarios para acesso ao sistema .",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Edward"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "edward@teste.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conta Criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Edward"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "edward@teste.com"
                                                },
                                                "privilege_id": {
                                                    "description": "Por Padrao todos os usuarios terão o privilegio de 3, que significa cliente, no momento o valor apresentado sera como null, mas posteriormente ele é alterado",
                                                    "type": "string",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos obrigatórios não foram preenchidos. Caso um ou mais campos sejam omitidos, será exibida uma mensagem de erro.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de erro quando um campo obrigatório não é enviado.",
                                            "type": "string",
                                            "example": "O campo nome é Obrigatorio (and 1 more error)"
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "Se o campo for emitido, sera exibido o erro no campo especifico.",
                                                        "type": "string",
                                                        "example": "O campo Name é Obrigatorio"
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "Caso, o email que o usuario esteja tentando cadastrar ja exista.",
                                                        "type": "string",
                                                        "example": "Este email ja esta em uso!"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Adicione o token no cabeçalho Authorization no formato: Bearer <seu_token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}